docs(checkers): split modules, apply EXAMPLE_STANDARD, pin cougr-core 1.1.0
Splits the monolithic lib.rs into components.rs (board/turn/status
component types and the small fixed-capacity collection) and
systems.rs (move/capture legality, promotion, win detection) per
EXAMPLE_STANDARD.md §3. Rewrites README.md to cover all eight §4
sections, marks the example transitional (it predates cougr-core's
component conventions entirely), bumps cougr-core to 1.1.0, and
commits Cargo.lock. No game rule logic was changed.

Refs #223
